requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-jazzmin
drf-yasg 